+++ Silicon/CometlakePkg/Include/RegAccess.h
@@ -28,15 +28,15 @@
 #include <PchReservedResources.h>
 #include <Register/PchRegsSpi.h>
 #include <Register/PchRegsGpio.h>
-#include <Register/PchRegsGpioCnl.h>
+#include <Register/PchRegsGpioCml.h>
 #include <Register/PchRegsSerialIo.h>
-#include <Register/PchRegsSerialIoCnl.h>
+#include <Register/PchRegsSerialIoCml.h>
 #include <Register/PchRegsLpc.h>
-#include <Register/PchRegsLpcCnl.h>
+#include <Register/PchRegsLpcCml.h>
 #include <Register/PchRegsP2sb.h>
 #include <Register/PchRegsPcr.h>
 #include <Register/PchRegsPmc.h>
-#include <Register/PchRegsPmcCnl.h>
+#include <Register/PchRegsPmcCml.h>
 #include <Register/PchRegsPcie.h>
 #include <Register/PchRegsDmi.h>
 #include <Register/PchRegsScsSd.h>
